feat: validator layer, shell fixes, slog migration, config persistence, User.ID int64 (Tasks 134,135,141,143,144,104)
Task 134: Created validation middleware with go-playground/validator/v10
Task 135: Fixed CreateInitialVersion (stores fullText), checkAndRevokeIfNeeded (revocation logic), ChatStream (persists conversations/messages)
Task 141: Replaced fmt.Printf with slog in logger middleware, added RequestID middleware
Task 143: Added config SaveOverride/LoadOverride with YAML file persistence
Task 144: Changed User.ID from uint to int64, updated all related models/repos/services
Task 104: CSRF middleware already implemented and mounted (verified)

+++ backend/internal/model/user.go
@@ -5,7 +5,7 @@ import (
 )
 
 type User struct {
-	ID              uint      `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID              int64      `gorm:"primaryKey;autoIncrement" json:"id"`
 	Email           string    `gorm:"uniqueIndex;not null;size:255" json:"email"`
 	PasswordHash    string    `gorm:"not null;size:255" json:"-"`
 	Username        string    `gorm:"uniqueIndex;not null;size:64" json:"username"`
@@ -27,7 +27,7 @@ func (User) TableName() string {
 
 type JudgeQualification struct {
 	ID          uint       `gorm:"primaryKey;autoIncrement" json:"id"`
-	UserID      uint       `gorm:"not null;index" json:"user_id"`
+	UserID      int64       `gorm:"not null;index" json:"user_id"`
 	ContentType string     `gorm:"not null;size:50" json:"content_type"`
 	QualifiedAt time.Time  `gorm:"not null;default:now()" json:"qualified_at"`
 	RevokedAt   *time.Time `json:"revoked_at,omitempty"`
@@ -40,10 +40,10 @@ func (JudgeQualification) TableName() string {
 
 type ReputationLog struct {
 	ID        uint      `gorm:"primaryKey;autoIncrement" json:"id"`
-	UserID    uint      `gorm:"not null;index" json:"user_id"`
+	UserID    int64      `gorm:"not null;index" json:"user_id"`
 	Delta     int       `gorm:"not null" json:"delta"`
 	Reason    string    `gorm:"not null;size:100" json:"reason"`
-	RelatedID *uint     `json:"related_id,omitempty"`
+	RelatedID *int64     `json:"related_id,omitempty"`
 	CreatedAt time.Time `json:"created_at"`
 }
 
@@ -53,7 +53,7 @@ func (ReputationLog) TableName() string {
 
 type OAuthAccount struct {
 	ID          uint      `gorm:"primaryKey;autoIncrement" json:"id"`
-	UserID      uint      `gorm:"not null;index" json:"user_id"`
+	UserID      int64      `gorm:"not null;index" json:"user_id"`
 	Provider    string    `gorm:"not null;size:20" json:"provider"`
 	ProviderUID string    `gorm:"not null;size:255" json:"provider_uid"`
 	AccessToken string    `gorm:"type:text" json:"-"`
